requirements.txt
streamlit
torch
torchvision
timm
opencv-python
numpy
pillow